refactor: 사용하지 않는 컴포넌트 삭제 및 코드 분할
