--- frontend/src/views/About.vue
@@ -1,20 +1,0 @@
-<script setup>
-
-</script>
-
-<script>
-  export default {
-    name: 'About',
-  };
-</script>
-
-<template>
-  <div>
-    <h1>About Page</h1>
-    <p>This is the about page!</p>
-  </div>
-</template>
-
-<style scoped>
-
-</style>
