+++ frontend/src/views/Account.vue
@@ -1,0 +1,13 @@
+<script setup>
+
+</script>
+
+<template>
+    <div>
+        회원가입
+    </div>
+</template>
+
+<style scoped>
+
+</style>
